refactor: restructured createNote command and related utilities
