+++ src/command/command/index.mts
@@ -1,3 +1,4 @@
 import collectAdmonitions from "./collectAdmonitions/index.mjs";
+import createNote from "./createNote/index.mjs";
 
-export { collectAdmonitions };
+export { collectAdmonitions, createNote };
