+++ src/command/lib/fs/index.mts
@@ -31,3 +31,12 @@ export function getWorkspaceFoldersOrShowError() {
   }
   return workspaceFolders;
 }
+
+export async function isFileExists(uri: vscode.Uri): Promise<boolean> {
+  try {
+    await vscode.workspace.fs.stat(uri);
+    return true;
+  } catch {
+    return false;
+  }
+}
